Update Ansible config
This was originally an attempt to use Python 3.9 on the server, but
there are multiple issus with that, including that the apt uwsgi package
doesn't work with Python 3.9.

So, I reverted back to 3.8 but kept some of the related changes and
added a couple other improvements, in particular using `current` rather
than a specific version number in config files.

Original message:

I was only using Python 3.8 because that seems to be the default Python
3 on Ubuntu 20.04, but Python 3.9 is also available. The issue with this
is that some apt packages, like psycopg2, can only be installed under
Python 3.8, but that doesn't work when you tell Ansible to use Python
3.9. There might be a better way to handle all this, but this solution
actually requires installing less stuff on the server, which is good.